requirements.txt
accelerate==0.32.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
bert-score==0.3.13
causal_conv1d==1.5.3.post1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
-e git+https://github.com/AndrasFerenczy/thinking_cobra.git@dd062739bd782d5ea1b21b3a0ae4bc895116830d#egg=cobra_vlm
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
datasets==4.4.1
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
draccus @ git+https://github.com/dlwh/draccus@28df342def31bc5e9b8560ab7ecf51fb37dcd4a6
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.24.0
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jsonlines==4.0.0
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
mamba-ssm==1.2.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mergedeep==1.3.4
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycocoevalcap==1.2
pycocotools==2.0.10
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safetensors==0.6.2
sentry-sdk==2.43.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
timm==0.9.10
tokenizers==0.14.1
toml==0.10.2
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.34.1
triton==2.1.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.22.3
wcwidth==0.2.14
xxhash==3.6.0
yarl==1.22.0
# External model API clients (optional, for testing against GPT-5, Gemini, Claude, Llama)
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
